Create I currently have currency to trade in a FTXplatform I'm ready to do a 30 sec. Trade but I can't read the data and predict the trend to minimize the risk
I'm here to learn how to read the market or the block chain to predict the trend to minimize the risk of a large quick trade. Next is to build on that information and profit from each trade. I am hoping that this site can teach me that. I also have money in a crypto wallet in crypto.com.